LG_Backup/ConvAI/ConvAI/apps/ker/requirements.txt
allennlp==1.2.0
allennlp-models==1.2.0
Flask==1.1.2
fuzzywuzzy==0.18.0
gensim==3.8.3
h5py==2.10.0
Keras==2.2.4
lxml==4.5.1
matplotlib==3.2.1
neo4j==4.1.1
nltk==3.5
numpy==1.18.4
openpyxl==3.0.3
pandas==1.1.5
py-rouge==1.1
py2neo==2020.0.0
pyhocon==0.3.55
pytest==5.4.3
python-docx==0.8.10
rdflib==5.0.0
rdflib-jsonld==0.5.0
scikit-learn==0.24.1
scipy==1.4.1
sentence-transformers==0.3.9
six==1.15.0
spacy==2.2.4
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-hub==0.7.0
torch==1.7.0 -f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.8.1
torchaudio==0.7.0
tqdm==4.46.0
transformers==3.4.0
urllib3==1.24.3
Werkzeug==1.0.1
xlrd==1.2.0
spacy_langdetect==0.1.2
